+++ packages/stream_video_flutter/lib/src/call_participants/screen_share_call_participants_content.dart
@@ -37,7 +37,7 @@ class ScreenShareCallParticipantsContent extends StatelessWidget {
     Call call,
     CallParticipantState participant,
   ) {
-    return StreamCallParticipant(
+    return StreamParticipantTile(
       key: ValueKey(
         '${participant.uniqueParticipantKey}-screenshare-video',
       ),